DRY to_dict
