+++ posthog/models/filters/stickiness_filter.py
@@ -10,7 +10,7 @@ from django.utils import timezone
 from posthog.constants import INTERVAL, STICKINESS_DAYS
 from posthog.models.entity import Entity
 from posthog.models.event import Event
-from posthog.models.filters.base_filter import BaseFilter
+from posthog.models.filters.base_filter import BaseFilter, SerializerWithDateMixin
 from posthog.models.filters.filter import Filter
 from posthog.models.filters.mixins.common import CompareMixin, DateMixin, IntervalMixin, OffsetMixin, ShownAsMixin
 from posthog.models.filters.mixins.property import PropertyMixin
@@ -35,7 +35,7 @@ class StickinessFilter(
     OffsetMixin,
     CompareMixin,
     ShownAsMixin,
-    DateMixin,
+    SerializerWithDateMixin,
     BaseFilter,
 ):
     get_earliest_timestamp: Callable
@@ -51,47 +51,6 @@ class StickinessFilter(
         if not get_earliest_timestamp:
             raise ValueError("Callable must be provided when date filtering is all time")
         self.get_earliest_timestamp = get_earliest_timestamp  # type: ignore
-
-    def to_dict(self) -> Dict[str, Any]:
-        ret = {}
-
-        for key in dir(self):
-            value = getattr(self, key)
-            if key in [
-                "entities",
-                "determine_time_delta",
-                "date_filter_Q",
-                "custom_date_filter_Q",
-                "properties_to_Q",
-                "toJSON",
-                "to_dict",
-            ] or key.startswith("_"):
-                continue
-            if isinstance(value, list) and len(value) == 0:
-                continue
-            if not isinstance(value, list) and not value:
-                continue
-            if key == "date_from" and not self._date_from:
-                continue
-            if key == "date_to" and not self._date_to:
-                continue
-            if isinstance(value, datetime):
-                value = value.isoformat()
-            if not isinstance(value, (list, bool, int, float, str)):
-                # Try to see if this object is json serializable
-                try:
-                    json.dumps(value)
-                except:
-                    continue
-            if isinstance(value, Entity):
-                value = value.to_dict()
-            if key == "properties" and isinstance(value, list) and isinstance(value[0], Property):
-                value = [prop.to_dict() for prop in value]
-            if isinstance(value, list) and isinstance(value[0], Entity):
-                value = [entity.to_dict() for entity in value]
-            ret[key] = value
-
-        return ret
 
     def trunc_func(self, field_name: str) -> Union[TruncMinute, TruncHour, TruncDay, TruncWeek, TruncMonth]:
         if self.interval == "minute":
